+++ lib/lerk/hints.rb
@@ -10,6 +10,8 @@ module Lerk
   module Hints
     EVENT_KEY_HINTS = 'cmd_hints_total'
 
+    TAGS_SORT_ORDER = { '--name' => :name, '--hints' => :hints, nil => :name }
+
     def self.register(bot)
       @bot = bot
       @logger = ::Lerk.logger
@@ -60,6 +62,32 @@ module Lerk
       ) do |event, arg|
         command_tiplist(event, arg)
       end
+
+      @bot.command(
+        [:tags],
+        description: 'List all known hint tags',
+        usage: '!tags [--name|--hints]',
+        min_args: 0,
+        max_args: 1,
+      ) do |event, sort_order|
+        sort_order = TAGS_SORT_ORDER[sort_order]
+        if sort_order
+          command_tags(event, sort_order)
+        else
+          'Invalid sort order, check command usage.'
+        end
+      end
+
+      @bot.command(
+        :reloadhints,
+        description: 'Reload hints',
+        usage: '!reloadhints',
+        min_args: 0,
+        max_args: 0,
+        permission_level: Lerk::PERMISSION_LEVEL_HINTS_ADMIN,
+      ) do  |event|
+        command_reloadhints(event)
+      end
     end
 
     def self.command_hint(event, group: nil, tag: nil)
@@ -84,6 +112,8 @@ module Lerk
     end
 
     def self.command_tiplist(event, arg)
+      @logger.command(event, 'tiplist', { category: arg })
+
       if arg.nil?
         'Check Sticky Messages in #newbie-corner in the official NS2 Discord!'
       elsif arg == 'guide'
@@ -95,6 +125,45 @@ module Lerk
       else
         "Invalid argument. Pick one of: (guide, movement, voyeur)"
       end
+    end
+
+    def self.command_tags(event, sort_order)
+      @logger.command(event, 'tags', { sort_order: sort_order })
+
+      data = HintTag.map do |tag|
+        [
+          tag.tag,
+          tag.hints_dataset.count
+        ]
+      end
+
+      if sort_order == :name
+        data = data.sort_by(&:first)
+      elsif sort_order == :hints
+        data = data.sort_by(&:last).reverse
+      end
+
+      data.
+        map { |ary| "#{ ary.first }: #{ ary.last }" }.
+        join("\n")
+    end
+
+    def self.command_reloadhints(event)
+      discord_user = Util.discord_user_from_database(event)
+
+      @logger.command(event, 'reloadhints')
+
+      puts "\n\n#{ event.class }\n\n"
+
+      event.send_message "Reloading hints.\nStarting download..."
+      source = Hints::HTTPDownloadSource.new
+      parser = Hints::Parser.new(source: source)
+      hints = parser.parse
+      event.send_message "Read #{ hints.length } hints from source."
+      exporter = Hints::SequelExporter.new(hints)
+      exporter.export
+      event << "Stored #{ Hint.count } hints in database."
+      event << 'Have a nice day!'
     end
   end
 end